+++ src/main/java/dk/os2opgavefordeler/service/EmploymentService.java
@@ -7,8 +7,8 @@ import java.util.List;
 import java.util.Optional;
 
 public interface EmploymentService {
-	Optional<Employment> getEmployment(int id);
-	Optional<EmploymentPO> getEmploymentPO(int id);
+	Optional<Employment> getEmployment(long id);
+	Optional<EmploymentPO> getEmploymentPO(long id);
 	List<Employment> findByEmail(String email);
 
 	List<Employment> getAll();
